pyproject.toml
[project]
name = "fast-bitrix24-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.8.0",
    "langchain-mcp-adapters>=0.1.7",
    "langchain[openai]>=0.3.25",
    "langgraph>=0.4.8",
    "loguru>=0.7.3",
    "orm-bitrix24>=0.1.4",
    "python-dotenv>=1.1.0",
]


[[tool.uv.index]]
name = "testpypi"
url = "https://test.pypi.org/simple/"
publish-url = "https://test.pypi.org/legacy/"
explicit = true


[project.urls]
Homepage = "https://github.com/darkClaw921/fast-bitrix24-mcp"
Documentation = "https://github.com/darkClaw921/fast-bitrix24-mcp"
Repository = "https://github.com/darkClaw921/fast-bitrix24-mcp.git"
"Bug Tracker" = "https://github.com/darkClaw921/fast-bitrix24-mcp/issues"
Changelog = "https://github.com/darkClaw921/fast-bitrix24-mcp/blob/master/CHANGELOG.md"
